Move hsa_activity_trace_entry_t to roctracer_trace_entries.h
